# perf(ts): denoise baseline mem usage